Conflicts:
	README.md
	build.xml
	build/doxygen/doxygen_footer.html
	build/doxygen/doxygen_header.html
	dev/codesniffer/README
	dev/dolibarr_changes.txt
	dev/examples/create_product.php
	dev/examples/get_contracts.php
	dev/initdata/generate-commande.php
	dev/initdata/generate-facture.php
	dev/initdata/generate-produit.php
	dev/initdata/generate-propale.php
	dev/ldap/ldapadd_sample1.txt
	dev/ldap/ldapadd_sample2.txt
	dev/ldap/ldapadd_sample3.txt
	dev/ldap/ldapdelete_sample1.txt
	dev/ldap/ldapdelete_sample3.txt
	dev/ldap/ldapmodify_sample1.txt
	dev/translation/autotranslator.class.php
	htdocs/.gitignore
	htdocs/comm/action/card.php
	htdocs/comm/action/class/actioncomm.class.php
	htdocs/comm/card.php
	htdocs/comm/list.php
	htdocs/comm/propal/class/propalestats.class.php
	htdocs/commande/apercu.php
	htdocs/commande/class/commande.class.php
	htdocs/compta/facture.php
	htdocs/compta/facture/class/facturestats.class.php
	htdocs/contact/class/contact.class.php
	htdocs/contact/list.php
	htdocs/contrat/class/contrat.class.php
	htdocs/contrat/list.php
	htdocs/core/class/CMailFile.class.php
	htdocs/core/class/commonobject.class.php
	htdocs/core/class/hookmanager.class.php
	htdocs/core/class/html.form.class.php
	htdocs/core/class/html.formactions.class.php
	htdocs/core/class/html.formsms.class.php
	htdocs/core/db/mysqli.class.php
	htdocs/core/lib/functions2.lib.php
	htdocs/core/lib/security2.lib.php
	htdocs/core/menus/standard/eldy.lib.php
	htdocs/core/tpl/contacts.tpl.php
	htdocs/core/tpl/login.tpl.php
	htdocs/fichinter/class/fichinter.class.php
	htdocs/holiday/admin/holiday.php
	htdocs/imports/import.php
	htdocs/includes/fpdfi/fpdi.php
	htdocs/includes/fpdfi/fpdi2tcpdf_bridge.php
	htdocs/includes/jquery/css/redmond/images/ui-bg_flat_55_fbec88_40x100.png
	htdocs/includes/jquery/css/redmond/images/ui-bg_gloss-wave_55_5c9ccc_500x100.png
	htdocs/includes/jquery/css/redmond/images/ui-icons_d8e7f3_256x240.png
	htdocs/includes/jquery/css/redmond/jquery-ui-latest.custom.css
	htdocs/includes/jquery/css/smoothness/jquery-ui-latest.custom.css
	htdocs/includes/jquery/css/ui-lightness/images/ui-icons_ffd27a_256x240.png
	htdocs/includes/jquery/css/ui-lightness/jquery-ui-latest.custom.css
	htdocs/includes/jquery/js/jquery-latest.min.js
	htdocs/includes/jquery/js/jquery-ui-latest.custom.min.js
	htdocs/includes/jquery/plugins/datatables/css/jquery.dataTables_jui.css
	htdocs/includes/jquery/plugins/datatables/extras/TableTools/swf/copy_csv_xls.swf
	htdocs/includes/jquery/plugins/datatables/extras/TableTools/swf/copy_csv_xls_pdf.swf
	htdocs/includes/jquery/plugins/datatables/license-gpl2.txt
	htdocs/includes/jquery/plugins/layout/jquery.layout-latest.js
	htdocs/includes/phpexcel/PHPExcel/CachedObjectStorageFactory.php
	htdocs/includes/phpexcel/PHPExcel/Settings.php
	htdocs/includes/phpexcel/PHPExcel/Writer/Excel5/Worksheet.php
	htdocs/install/doctemplates/thirdparties/template_thirdparty.odt
	htdocs/install/mssql/functions/functions.sql
	htdocs/langs/en_US/agenda.lang
	htdocs/langs/en_US/categories.lang
	htdocs/langs/en_US/companies.lang
	htdocs/langs/en_US/suppliers.lang
	htdocs/langs/en_US/trips.lang
	htdocs/langs/fr_FR/main.lang
	htdocs/main.inc.php
	htdocs/product/class/product.class.php
	htdocs/product/index.php
	htdocs/product/stock/class/mouvementstock.class.php
	htdocs/product/stock/replenish.php
	htdocs/projet/tasks/time.php
	htdocs/public/demo/index.php
	htdocs/societe/class/societe.class.php
	htdocs/societe/soc.php
	htdocs/societe/societe.php
	htdocs/theme/auguria/img/stcomm-1.png
	htdocs/theme/auguria/img/stcomm-1_grayed.png
	htdocs/theme/auguria/img/stcomm0.png
	htdocs/theme/auguria/img/stcomm0_grayed.png
	htdocs/theme/auguria/img/stcomm1_grayed.png
	htdocs/theme/auguria/img/stcomm2_grayed.png
	htdocs/theme/auguria/img/stcomm3_grayed.png
	htdocs/theme/auguria/img/stcomm4_grayed.png
	htdocs/theme/bureau2crea/img/stcomm-1_grayed.png
	htdocs/theme/bureau2crea/img/stcomm0.png
	htdocs/theme/bureau2crea/img/stcomm0_grayed.png
	htdocs/theme/bureau2crea/img/stcomm1_grayed.png
	htdocs/theme/bureau2crea/img/stcomm2_grayed.png
	htdocs/theme/bureau2crea/img/stcomm3_grayed.png
	htdocs/theme/bureau2crea/img/stcomm4_grayed.png
	htdocs/theme/cameleo/img/stcomm0.png
	htdocs/theme/common/nophoto.jpg
	htdocs/theme/eldy/img/delete.png
	htdocs/theme/eldy/img/help.png
	htdocs/theme/eldy/img/info.png
	htdocs/theme/eldy/img/object_address.png
	htdocs/theme/eldy/img/object_phoning.png
	htdocs/theme/eldy/img/searchclear.png
	htdocs/theme/eldy/img/sort_asc.png
	htdocs/theme/eldy/img/sort_asc_disabled.png
	htdocs/theme/eldy/img/sort_desc.png
	htdocs/theme/eldy/img/stcomm-1.png
	htdocs/theme/eldy/img/stcomm-1_grayed.png
	htdocs/theme/eldy/img/stcomm0.png
	htdocs/theme/eldy/img/stcomm0_grayed.png
	htdocs/theme/eldy/img/stcomm1_grayed.png
	htdocs/theme/eldy/img/stcomm2_grayed.png
	htdocs/theme/eldy/img/stcomm3_grayed.png
	htdocs/theme/eldy/img/stcomm4_grayed.png
	htdocs/user/class/user.class.php
	scripts/company/sync_contacts_dolibarr2ldap.php
	scripts/members/sync_members_dolibarr2ldap.php
	test/README
	test/phpunit/AdminLibTest.php
	test/phpunit/BankAccountTest.php
	test/phpunit/CMailFileTest.php
	test/phpunit/ChargeSocialesTest.php
	test/phpunit/CommandeTest.php
	test/phpunit/CommonObjectTest.php
	test/phpunit/CompanyBankAccountTest.php
	test/phpunit/ContratTest.php
	test/phpunit/DiscountTest.php
	test/phpunit/ExportTest.php
	test/phpunit/FactureFournisseurTest.php
	test/phpunit/FactureTest.php
	test/phpunit/FactureTestRounding.php
	test/phpunit/ImagesLibTest.php
	test/phpunit/ImportTest.php
	test/phpunit/PgsqlTest.php
	test/phpunit/ProductTest.php
	test/phpunit/PropalTest.php
	test/phpunit/SecurityTest.php
	test/phpunit/UserGroupTest.php

+++ htdocs/holiday/card.php
@@ -1,6 +1,6 @@
 <?php
 /* Copyright (C) 2011	   Dimitri Mouillard	<dmouillard@teclib.com>
- * Copyright (C) 2012-2013 Laurent Destailleur	<eldy@users.sourceforge.net>
+ * Copyright (C) 2012-2015 Laurent Destailleur	<eldy@users.sourceforge.net>
  * Copyright (C) 2012	   Regis Houssin		<regis.houssin@capnetworks.com>
  * Copyright (C) 2013	   Juanjo Menent		<jmenent@2byte.es>
  * Copyright (C) 2014	   Ferran Marcet		<fmarcet@2byte.es>
@@ -51,12 +51,12 @@ $now=dol_now();
  * Actions
  */
 
-// Si création de la demande
+// If create a request
 if ($action == 'create')
 {
 	$cp = new Holiday($db);
 
-    // Si pas le droit de créer une demande
+    // If no right to create a request
     if (($userid == $user->id && empty($user->rights->holiday->write)) || ($userid != $user->id && empty($user->rights->holiday->write_all)))
     {
     	$error++;
@@ -66,12 +66,15 @@ if ($action == 'create')
 
     if (! $error)
     {
+    	$db->begin();
+
 	    $date_debut = dol_mktime(0, 0, 0, GETPOST('date_debut_month'), GETPOST('date_debut_day'), GETPOST('date_debut_year'));
 	    $date_fin = dol_mktime(0, 0, 0, GETPOST('date_fin_month'), GETPOST('date_fin_day'), GETPOST('date_fin_year'));
 	    $date_debut_gmt = dol_mktime(0, 0, 0, GETPOST('date_debut_month'), GETPOST('date_debut_day'), GETPOST('date_debut_year'), 1);
 	    $date_fin_gmt = dol_mktime(0, 0, 0, GETPOST('date_fin_month'), GETPOST('date_fin_day'), GETPOST('date_fin_year'), 1);
 	    $starthalfday=GETPOST('starthalfday');
 	    $endhalfday=GETPOST('endhalfday');
+	    $type=GETPOST('type');
 	    $halfday=0;
 	    if ($starthalfday == 'afternoon' && $endhalfday == 'morning') $halfday=2;
 	    else if ($starthalfday == 'afternoon') $halfday=-1;
@@ -81,21 +84,21 @@ if ($action == 'create')
 	    $description = trim(GETPOST('description'));
 	    $userID = GETPOST('userID');
 
-	    // Si pas de date de début
+	    // If no start date
 	    if (empty($date_debut))
 	    {
 	        header('Location: card.php?action=request&error=nodatedebut');
 	        exit;
 	    }
 
-	    // Si pas de date de fin
+	    // If no end date
 	    if (empty($date_fin))
 	    {
 	        header('Location: card.php?action=request&error=nodatefin');
 	        exit;
 	    }
 
-	    // Si date de début après la date de fin
+	    // If start date after end date
 	    if ($date_debut > $date_fin)
 	    {
 	        header('Location: card.php?action=request&error=datefin');
@@ -110,7 +113,7 @@ if ($action == 'create')
 	        exit;
 	    }
 
-	    // Si aucun jours ouvrés dans la demande
+	    // If there is no Business Days within request 
 	    $nbopenedday=num_open_day($date_debut_gmt, $date_fin_gmt, 0, 1, $halfday);
 	    if($nbopenedday < 0.5)
 	    {
@@ -118,7 +121,7 @@ if ($action == 'create')
 	        exit;
 	    }
 
-	    // Si pas de validateur choisi
+	    // If no validator designated
 	    if ($valideur < 1)
 	    {
 	        header('Location: card.php?action=request&error=Valideur');
@@ -131,18 +134,23 @@ if ($action == 'create')
 	    $cp->date_fin = $date_fin;
 	    $cp->fk_validator = $valideur;
 		$cp->halfday = $halfday;
+		$cp->fk_type = $type;
 
-	    $verif = $cp->create($userid);
+		$verif = $cp->create($user);
 
-	    // Si pas d'erreur SQL on redirige vers la fiche de la demande
+	    // If no SQL error we redirect to the request card
 	    if ($verif > 0)
 	    {
-	        header('Location: card.php?id='.$verif);
+			$db->commit();
+
+	    	header('Location: card.php?id='.$verif);
 	        exit;
 	    }
 	    else
-	    {
-	        // Sinon on affiche le formulaire de demande avec le message d'erreur SQL
+		{
+	    	$db->rollback();
+
+	        // Otherwise we display the request form with the SQL error message
 	        header('Location: card.php?action=request&error=SQL_Create&msg='.$cp->error);
 	        exit;
 	    }
@@ -162,7 +170,7 @@ if ($action == 'update')
 	else if ($starthalfday == 'afternoon') $halfday=-1;
 	else if ($endhalfday == 'morning') $halfday=1;
 
-    // Si pas le droit de modifier une demande
+    // If no right to modify a request 
     if (! $user->rights->holiday->write)
     {
         header('Location: card.php?action=request&error=CantUpdate');
@@ -174,40 +182,40 @@ if ($action == 'update')
 
 	$canedit=(($user->id == $cp->fk_user && $user->rights->holiday->write) || ($user->id != $cp->fk_user && $user->rights->holiday->write_all));
 
-	// Si en attente de validation
+	// If under validation
     if ($cp->statut == 1)
     {
-        // Si c'est le créateur ou qu'il a le droit de tout lire / modifier
+        // If this is the requestor or has read/write rights
         if ($canedit)
         {
             $valideur = $_POST['valideur'];
             $description = trim($_POST['description']);
 
-            // Si pas de date de début
+            // If no start date
             if (empty($_POST['date_debut_'])) {
                 header('Location: card.php?id='.$_POST['holiday_id'].'&action=edit&error=nodatedebut');
                 exit;
             }
 
-            // Si pas de date de fin
+            // If no end date
             if (empty($_POST['date_fin_'])) {
                 header('Location: card.php?id='.$_POST['holiday_id'].'&action=edit&error=nodatefin');
                 exit;
             }
 
-            // Si date de début après la date de fin
+            // If start date after end date
             if ($date_debut > $date_fin) {
                 header('Location: card.php?id='.$_POST['holiday_id'].'&action=edit&error=datefin');
                 exit;
             }
 
-            // Si pas de valideur choisi
+            // If no validator designated
             if ($valideur < 1) {
                 header('Location: card.php?id='.$_POST['holiday_id'].'&action=edit&error=Valideur');
                 exit;
             }
 
-            // Si pas de jours ouvrés dans la demande
+            // If there is no Business Days within request
             $nbopenedday=num_open_day($date_debut_gmt, $date_fin_gmt, 0, 1, $halfday);
             if ($nbopenedday < 0.5)
             {
@@ -230,7 +238,7 @@ if ($action == 'update')
             }
             else
            {
-                // Sinon on affiche le formulaire de demande avec le message d'erreur SQL
+                // Otherwise we display the request form with the SQL error message
                 header('Location: card.php?id='.$_POST['holiday_id'].'&action=edit&error=SQL_Create&msg='.$cp->error);
                 exit;
             }
@@ -241,45 +249,42 @@ if ($action == 'update')
     }
 }
 
-// Si suppression de la demande
-if ($action == 'confirm_delete' && GETPOST('confirm') == 'yes')
+// If delete of request
+if ($action == 'confirm_delete' && GETPOST('confirm') == 'yes' && $user->rights->holiday->delete)
 {
-    if($user->rights->holiday->delete)
-    {
-    	$error=0;
+	$error=0;
 
-    	$db->begin();
+	$db->begin();
 
-        $cp = new Holiday($db);
-        $cp->fetch($id);
+	$cp = new Holiday($db);
+	$cp->fetch($id);
 
-        $canedit=(($user->id == $cp->fk_user && $user->rights->holiday->write) || ($user->id != $cp->fk_user && $user->rights->holiday->write_all));
+	$canedit=(($user->id == $cp->fk_user && $user->rights->holiday->write) || ($user->id != $cp->fk_user && $user->rights->holiday->write_all));
 
-        // Si c'est bien un brouillon
-        if ($cp->statut == 1 || $cp->statut == 3)
-        {
-            // Si l'utilisateur à le droit de lire cette demande, il peut la supprimer
-            if ($canedit)
-            {
-                $result=$cp->delete($id);
-            }
-            else
-            {
-                $error = $langs->trans('ErrorCantDeleteCP');
-            }
-        }
+    // If this is a rough draft
+	if ($cp->statut == 1 || $cp->statut == 3)
+	{
+		// Si l'utilisateur à le droit de lire cette demande, il peut la supprimer
+		if ($canedit)
+		{
+			$result=$cp->delete($id);
+		}
+		else
+		{
+			$error = $langs->trans('ErrorCantDeleteCP');
+		}
+	}
 
-       	if (! $error)
-        {
-          	$db->commit();
-           	header('Location: index.php');
-           	exit;
-        }
-        else
-        {
-        	$db->rollback();
-        }
-    }
+	if (! $error)
+	{
+		$db->commit();
+		header('Location: list.php');
+		exit;
+	}
+	else
+	{
+		$db->rollback();
+	}
 }
 
 // Si envoi de la demande
@@ -381,7 +386,7 @@ if ($action == 'confirm_send')
 
 
 // Si Validation de la demande
-if($action == 'confirm_valid')
+if ($action == 'confirm_valid')
 {
     $cp = new Holiday($db);
     $cp->fetch($id);
@@ -466,7 +471,7 @@ if($action == 'confirm_valid')
 
 if ($action == 'confirm_refuse')
 {
-    if (!empty($_POST['detail_refuse']))
+    if (! empty($_POST['detail_refuse']))
     {
         $cp = new Holiday($db);
         $cp->fetch($_GET['id']);
@@ -673,7 +678,7 @@ if (empty($id) || $action == 'add' || $action == 'request' || $action == 'create
     else
     {
         // Formulaire de demande de congés payés
-        print_fiche_titre($langs->trans('MenuAddCP'));
+        print_fiche_titre($langs->trans('MenuAddCP'), '', 'title_hrm.png');
 
         // Si il y a une erreur
         if (GETPOST('error')) {
@@ -747,11 +752,31 @@ if (empty($id) || $action == 'add' || $action == 'request' || $action == 'create
         print '<form method="POST" action="'.$_SERVER['PHP_SELF'].'" onsubmit="return valider()" name="demandeCP">'."\n";
         print '<input type="hidden" name="action" value="create" />'."\n";
         print '<input type="hidden" name="userID" value="'.$userid.'" />'."\n";
-        print '<div class="tabBar">';
-        print '<span>'.$langs->trans('DelayToRequestCP',$cp->getConfCP('delayForRequest')).'</span><br /><br />';
+
+        dol_fiche_head();
+
+        $out='';
+        $typeleaves=$cp->getTypes(1,1);
+    	foreach($typeleaves as $key => $val)
+		{
+			$nb_type = $cp->getCPforUser($user->id, $val['rowid']);
+			$nb_holiday += $nb_type;
+			$out .= ' - '.$val['label'].': <strong>'.($nb_type?price2num($nb_type):0).'</strong><br>';
+		}
+        print $langs->trans('SoldeCPUser', round($nb_holiday,5)).'<br>';
+		print $out;
+
+        dol_fiche_end();
+
+
+        dol_fiche_head();
+
+        //print '<span>'.$langs->trans('DelayToRequestCP',$cp->getConfCP('delayForRequest')).'</span><br /><br />';
 
         print '<table class="border" width="100%">';
         print '<tbody>';
+
+        // User
         print '<tr>';
         print '<td class="fieldrequired">'.$langs->trans("User").'</td>';
         print '<td>';
@@ -761,10 +786,26 @@ if (empty($id) || $action == 'add' || $action == 'request' || $action == 'create
         	print '<input type="hidden" name="userid" value="'.$userid.'">';
         }
         else print $form->select_users(GETPOST('userid')?GETPOST('userid'):$user->id,'userid',0,'',0);
-        $nb_holiday = $cp->getCPforUser($user->id) / $cp->getConfCP('nbHolidayDeducted');
-        print ' &nbsp; <span>'.$langs->trans('SoldeCPUser', round($nb_holiday,2)).'</span>';
         print '</td>';
         print '</tr>';
+
+        // Type
+        print '<tr>';
+        print '<td class="fieldrequired">'.$langs->trans("Type").'</td>';
+        print '<td>';
+        $typeleaves=$cp->getTypes(1,1);
+        $arraytypeleaves=array();
+        foreach($typeleaves as $key => $val)
+        {
+        	$labeltoshow = $val['label'];
+        	$labeltoshow .= ($val['delay'] > 0 ? ' ('.$langs->trans("NoticePeriod").': '.$val['delay'].' '.$langs->trans("days").')':'');
+			$arraytypeleaves[$val['rowid']]=$labeltoshow;
+        }
+        print $form->selectarray('type', $arraytypeleaves, (GETPOST('type')?GETPOST('type'):''), 1);
+        print '</td>';
+        print '</tr>';
+
+        // Date start
         print '<tr>';
         print '<td class="fieldrequired">'.$langs->trans("DateDebCP").' ('.$langs->trans("FirstDayOfHoliday").')</td>';
         print '<td>';
@@ -779,6 +820,8 @@ if (empty($id) || $action == 'add' || $action == 'request' || $action == 'create
         print $form->selectarray('starthalfday', $listhalfday, (GETPOST('starthalfday')?GETPOST('starthalfday'):'morning'));
         print '</td>';
         print '</tr>';
+
+        // Date end
         print '<tr>';
         print '<td class="fieldrequired">'.$langs->trans("DateFinCP").' ('.$langs->trans("LastDayOfHoliday").')</td>';
         print '<td>';
@@ -797,7 +840,6 @@ if (empty($id) || $action == 'add' || $action == 'request' || $action == 'create
         // Approved by
         print '<tr>';
         print '<td class="fieldrequired">'.$langs->trans("ReviewedByCP").'</td>';
-        // Liste des utiliseurs du groupe choisi dans la config
         $validator = new UserGroup($db);
         $excludefilter=$user->admin?'':'u.rowid <> '.$user->id;
         $valideurobjects = $validator->listUsersForGroup($excludefilter);
@@ -818,17 +860,17 @@ if (empty($id) || $action == 'add' || $action == 'request' || $action == 'create
 
         print '</tbody>';
         print '</table>';
-        print '<div style="clear: both;"></div>';
-        print '</div>';
-        print '</from>'."\n";
 
-        print '<center>';
+        dol_fiche_end();
+
+        print '<div class="center">';
         print '<input type="submit" value="'.$langs->trans("SendRequestCP").'" name="bouton" class="button">';
         print '&nbsp; &nbsp; ';
         print '<input type="button" value="'.$langs->trans("Cancel").'" class="button" onclick="history.go(-1)">';
-        print '</center>';
-    }
+        print '</div>';
 
+        print '</from>'."\n";
+    }
 }
 else
 {
@@ -936,9 +978,9 @@ else
                 if ($action == 'edit' && $cp->statut == 1)
                 {
                     $edit = true;
-                    print '<form method="post" action="'.$_SERVER['PHP_SELF'].'?id='.$_GET['id'].'">'."\n";
+                    print '<form method="post" action="'.$_SERVER['PHP_SELF'].'?id='.$id.'">'."\n";
                     print '<input type="hidden" name="action" value="update"/>'."\n";
-                    print '<input type="hidden" name="holiday_id" value="'.$_GET['id'].'" />'."\n";
+                    print '<input type="hidden" name="holiday_id" value="'.$id.'" />'."\n";
                 }
 
                 print '<table class="border" width="100%">';
@@ -957,6 +999,15 @@ else
         		print '<td>';
         		print $userRequest->getNomUrl(1);
         		print '</td></tr>';
+
+		        // Type
+		        print '<tr>';
+		        print '<td>'.$langs->trans("Type").'</td>';
+		        print '<td>';
+		        $typeleaves=$cp->getTypes(1,1);
+		        print $typeleaves[$cp->fk_type]['label'];
+		        print '</td>';
+		        print '</tr>';
 
 			    $starthalfday=($cp->halfday == -1 || $cp->halfday == 2)?'afternoon':'morning';
 			    $endhalfday=($cp->halfday == 1 || $cp->halfday == 2)?'morning':'afternoon';
@@ -1037,6 +1088,7 @@ else
                     print '<td><textarea name="description" class="flat" rows="'.ROWS_3.'" cols="70">'.$cp->description.'</textarea></td>';
                     print '</tr>';
                 }
+
                 print '</tbody>';
                 print '</table>'."\n";
 
@@ -1067,14 +1119,15 @@ else
                 } else {
                     print '<tr>';
                     print '<td width="50%">'.$langs->trans('ReviewedByCP').'</td>';
-                    // Liste des utiliseurs du groupes choisi dans la config
-                    $idGroupValid = $cp->getConfCP('userGroup');
 
-                    $validator = new UserGroup($db);
-                    $valideur = $validator->listUsersForGroup('',1);
+			        $validator = new UserGroup($db);
+			        $excludefilter=$user->admin?'':'u.rowid <> '.$user->id;
+			        $valideurobjects = $validator->listUsersForGroup($excludefilter);
+			        $valideurarray = array();
+			        foreach($valideurobjects as $val) $valideurarray[$val->id]=$val->id;
 
                     print '<td>';
-                    $form->select_users($cp->fk_validator,"valideur",1,"",0,$valideur,'');
+        			print $form->select_dolusers($user->fk_user, "valideur", 1, "", 0, $valideurarray);	// By default, hierarchical parent
                     print '</td>';
                     print '</tr>';
                 }
